Fixed signed/unsigned mismatch in boolean comparison.
git-svn-id: http://svn.osgeo.org/qgis/branches/Lib_Refactoring-branch@5666 c8812cc2-4d05-0410-92ff-de0c093fc19c

+++ src/core/qgsattributeaction.cpp
@@ -73,8 +73,10 @@ QgsAttributeAction::aIter QgsAttributeAction::retrieveAction(unsigned int index)
   if (index >= 0 && index < mActions.size())
   {
     a_iter = mActions.begin();
-    for (int i = 0; i < index; ++i, ++a_iter)
-      {}
+    for (unsigned int i = 0; i < index; ++i, ++a_iter)
+    {
+        // idle
+    }
   }
   return a_iter;
 }
@@ -98,12 +100,12 @@ QString QgsAttributeAction::expandAction(QString action, const std::vector<std::
   // for the actual substitutions.
 
   QString expanded_action;
-  if (clickedOnValue >= 0 && clickedOnValue < values.size())
+  if (clickedOnValue >= 0 && clickedOnValue < static_cast<int>(values.size()))
     expanded_action = action.replace("%%", values[clickedOnValue].second);
   else
     expanded_action = action;
 
-  for (int i = 0; i < values.size(); ++i)
+  for (unsigned int i = 0; i < values.size(); ++i)
     {
       QString to_replace = "%" + values[i].first;
       expanded_action = expanded_action.replace(to_replace, values[i].second);
@@ -138,7 +140,7 @@ bool QgsAttributeAction::readXML(QDomNode& layer_node)
   if (!aaNode.isNull())
   {
     QDomNodeList actionsettings = aaNode.childNodes();
-    for (int i = 0; i < actionsettings.length(); ++i)
+    for (unsigned int i = 0; i < actionsettings.length(); ++i)
     {
       QDomElement setting = actionsettings.item(i).toElement();
       int capture = setting.attributeNode("capture").value().toInt();
